+++ js/main.js
@@ -1,4 +1,4 @@
 import showScreen from './showScreen';
-import welcomeScreen from './main-welcome';
+import welcomeScreen from './screens/main-welcome';
 
 showScreen(welcomeScreen);
